Strongly separate `clash-protocols-base` and `clash-protocols`
`clash-protocols-base` should only contain code and definitions related to the circuit plugin.
This includes the `Circuit` definition and `Protocol` typeclass. Furthermore we include instances
for types imported from underlying clash libraries such as tuples, `Vec` and `Signal`.
