+++ clash-protocols/src/Protocols/Internal.hs
@@ -5,7 +5,9 @@
 {-# LANGUAGE RoleAnnotations #-}
 {-# LANGUAGE TypeFamilyDependencies #-}
 {-# LANGUAGE UndecidableInstances #-}
+#if !MIN_VERSION_clash_prelude(1, 8, 2)
 {-# OPTIONS_GHC -fno-warn-orphans #-}
+#endif
 
 -- TODO: Hide internal documentation
 -- {-# OPTIONS_HADDOCK hide #-}
@@ -16,22 +18,22 @@ Internal module to prevent hs-boot files (breaks Haddock)
 module Protocols.Internal (
   module Protocols.Internal,
   module Protocols.Internal.Classes,
+  module Protocols.Plugin,
 ) where
 
 import Control.DeepSeq (NFData)
 import Data.Hashable (Hashable)
 import Data.Maybe (fromMaybe)
 import Data.Proxy
-import GHC.Base (Any)
 import Prelude hiding (const, map)
 
 import qualified Clash.Explicit.Prelude as CE
-import Clash.Prelude (Signal, type (*), type (+))
+import Clash.Prelude (type (*), type (+))
 import qualified Clash.Prelude as C
 
-import Protocols.Cpp (maxTupleSize)
+import Protocols.Circuit
 import Protocols.Internal.Classes
-import Protocols.Internal.TH (protocolTupleInstances)
+import Protocols.Plugin
 
 import Control.Arrow ((***))
 import Data.Coerce (coerce)
@@ -52,42 +54,6 @@ newtype Ack = Ack Bool
 -- | Acknowledge. Used in circuit-notation plugin to drive ignore components.
 instance Default Ack where
   def = Ack True
-
-{- | Circuit protocol with /Signal dom a/ in its forward direction, and
-/()/ in its backward direction. Convenient for exposing protocol
-internals, or simply for undirectional streams.
-
-Note: 'CSignal' exists to work around [issue 760](https://github.com/clash-lang/clash-compiler/issues/760)
-      in Clash, where type families with 'Signal' on the LHS are broken.
--}
-data CSignal (dom :: CE.Domain) (a :: Type)
-
-type role CSignal nominal representational
-
-instance Protocol () where
-  type Fwd () = ()
-  type Bwd () = ()
-
-{- | __NB__: The documentation only shows instances up to /3/-tuples. By
-default, instances up to and including /12/-tuples will exist. If the flag
-@large-tuples@ is set instances up to the GHC imposed limit will exist. The
-GHC imposed limit is either 62 or 64 depending on the GHC version.
--}
-instance Protocol (a, b) where
-  type Fwd (a, b) = (Fwd a, Fwd b)
-  type Bwd (a, b) = (Bwd a, Bwd b)
-
--- Generate n-tuple instances, where n > 2
-protocolTupleInstances 3 maxTupleSize
-
-instance (C.KnownNat n) => Protocol (C.Vec n a) where
-  type Fwd (C.Vec n a) = C.Vec n (Fwd a)
-  type Bwd (C.Vec n a) = C.Vec n (Bwd a)
-
--- XXX: Type families with Signals on LHS are currently broken on Clash:
-instance Protocol (CSignal dom a) where
-  type Fwd (CSignal dom a) = Signal dom a
-  type Bwd (CSignal dom a) = Signal dom ()
 
 {- | Left-to-right circuit composition.
 
@@ -609,20 +575,6 @@ simulateCircuit fwds bwds circ =
   (bwdSig, fwdSig) =
     toSignals circ $
       (simToSigFwd (Proxy @a) fwds, simToSigBwd (Proxy @b) bwds)
-
-{- | Picked up by "Protocols.Plugin" to process protocol DSL. See
-"Protocols.Plugin" for more information.
--}
-circuit :: Any
-circuit =
-  error "'protocol' called: did you forget to enable \"Protocols.Plugin\"?"
-
-{- | Picked up by "Protocols.Plugin" to tie circuits together. See
-"Protocols.Plugin" for more information.
--}
-(-<) :: Any
-(-<) =
-  error "(-<) called: did you forget to enable \"Protocols.Plugin\"?"
 
 {- | Allows for optional data.
 Depending on the value of @keep@, the data can either be included or left out.
